Move getApp to app.jsx
